fazendo mudanças no app
