+++ src/app.ts
@@ -1,6 +1,6 @@
 import express, {Express} from 'express'
 import cors from 'cors'
-import getAllEstudantes from './endpoints/estudante';
+import { AddressInfo } from "net";
 
 const app: Express = express();
 
@@ -8,6 +8,14 @@ app.use(express.json());
 app.use(cors());
 
 // estudantes
-app.get("/estudantes", getAllEstudantes);
+
+const server = app.listen(process.env.PORT || 3003, () => {
+    if (server) {
+       const address = server.address() as AddressInfo;
+       console.log(`Server is running in http://localhost:${address.port}`);
+    } else {
+       console.error(`Failure upon starting server.`);
+    }
+});
 
 export default app
