+++ src/endpoints/estudante.ts
@@ -2,10 +2,8 @@ import { Request, Response } from "express";
 import selectAllEstudantes from "../data/estudante";
 
 const getAllEstudantes = async (req: Request, res: Response) => {
-
     try {
         const estudantes = await selectAllEstudantes()
-
         res.status(200).send({ estudantes: estudantes })
     }
     catch (error: any) {
