+++ src/index.ts
@@ -1,15 +1,7 @@
-import { AddressInfo } from "net";
 import app from "./app";
 import { addTurma } from "./endpoints/criarTurma";
+import getAllEstudantes from './endpoints/estudante';
 
+app.post("/turmas", addTurma);
+app.get("/estudantes", getAllEstudantes);
 
-app.post("/turmas", addTurma)
-
-const server = app.listen(process.env.PORT || 3003, () => {
-    if (server) {
-       const address = server.address() as AddressInfo;
-       console.log(`Server is running in http://localhost:${address.port}`);
-    } else {
-       console.error(`Failure upon starting server.`);
-    }
-});
